requirements.txt
pytorch
pandas 
tqdm
Pillow
opencv-python
torchvision